user handlers

+++ domain/entity/service.go
@@ -4,5 +4,5 @@ type Status struct {
 	User   int32 `json:"user"`
 	Forum  int32 `json:"forum"`
 	Thread int32 `json:"thread"`
-	Post   int32 `json:"post"`
+	Post   int64 `json:"post"`
 }
